Extract duplicate-finding function
Adding the function to util.misc module and then importing it for use
was causing a circular import. It's time to break up the misc.
