+++ qgreenland/test/util/test_misc.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from qgreenland.constants import TaskType
-from qgreenland.util import misc
+from qgreenland.util import layer as layer_util
 
 
 def test_final_layer_dir(raster_layer_node):
@@ -12,10 +12,10 @@ def test_final_layer_dir(raster_layer_node):
         / 'Example raster'
     )
 
-    actual = misc.get_final_layer_dir(raster_layer_node)
+    actual = layer_util.get_final_layer_dir(raster_layer_node)
 
     assert expected == actual
 
 
 def test_vector_or_raster_gdal_remote(online_layer_node):
-    assert misc.vector_or_raster(online_layer_node) == 'Raster'
+    assert layer_util.vector_or_raster(online_layer_node) == 'Raster'
