+++ qgreenland/util/qgis/metadata.py
@@ -9,7 +9,7 @@ from qgreenland.constants import (
     INPUT_DIR,
 )
 from qgreenland.models.config.layer import ConfigLayer
-from qgreenland.util.misc import datasource_dirname
+from qgreenland.util.layer import datasource_dirname
 from qgreenland.util.template import load_template
